+++ src/main/java/bgu/spl/mics/application/passiveObjects/Report.java
@@ -18,13 +18,13 @@ public class Report {
 	private List<String> agentsSerialNumbers;
 	private List<String> agentsNames;
 	private String gadgetName;
-	private int timeIssued;
-	private int QTime;
-	private int timeCreated;
+	private long timeIssued;
+	private long QTime;
+	private long timeCreated;
 
 	//Constructors
 	public Report(String missionName, int mId, int moneyPennyId, List<String> agentsSerialNumbers,
-				  List<String> agentsNames, String gadgetName, int timeIssued, int QTime, int timeCreated){
+				  List<String> agentsNames, String gadgetName, int timeIssued, long QTime, long timeCreated){
 		this.missionName = missionName;
 		this.mId = mId;
 		this.moneyPennyId = moneyPennyId;
@@ -142,21 +142,21 @@ public class Report {
 	/**
 	 * Retrieves the time-tick in which Q Received the GadgetAvailableEvent for that mission.
 	 */
-	public int getQTime() {
+	public long getQTime() {
 		return QTime;
 	}
 
 	/**
 	 * Sets the time-tick in which Q Received the GadgetAvailableEvent for that mission.
 	 */
-	public void setQTime(int qTime) {
+	public void setQTime(long qTime) {
 		QTime = qTime;
 	}
 
 	/**
 	 * Retrieves the time when the mission was sent by an Intelligence Publisher.
 	 */
-	public int getTimeIssued() {
+	public long getTimeIssued() {
 		return timeIssued;
 	}
 
@@ -170,14 +170,14 @@ public class Report {
 	/**
 	 * Retrieves the time-tick when the report has been created.
 	 */
-	public int getTimeCreated() {
+	public long getTimeCreated() {
 		return timeCreated;
 	}
 
 	/**
 	 * Sets the time-tick when the report has been created.
 	 */
-	public void setTimeCreated(int timeCreated) {
+	public void setTimeCreated(long timeCreated) {
 		this.timeCreated = timeCreated;
 	}
 }
